Move identity extensions to own class
Move the extension methods for adding identity to the service collection to their own class.
